+++ Tabs Sidebar.novaextension/CHANGELOG.md
@@ -1,3 +1,8 @@
+## v1.1.0
+
+- Automatically make shell scripts executable if they are not already
+- Fixes certain commands not working
+
 ## v1.0.0
 
 Initial release